simple_Agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "mcp>=1.13.1",
    "modelcontextprotocol>=0.1.0",
    "python-dotenv>=1.1.1",
]
